refactor: factored out BoxWidget
